+++ ChangeLog.md
@@ -2,7 +2,19 @@
 
 ## Unreleased
 
-* `PatchMapWithMove` supports moves with a patch.
+* Create `PatchMapWithPatchingMove` variant which supports moves with a patch.
+
+## 0.0.2.0
+
+* Consistently provide:
+
+   - `Wrapped` instances
+
+   - `*WithIndex` instances
+
+   - `un*` newtype unwrappers
+
+  for `PatchMap`, `PatchIntMap`, and `PatchMapWithMove`.
 
 ## 0.0.1.0